gcc build fix
svn:sync-xref-src-repo-rev: r80558

+++ src/VBox/Runtime/common/string/RTStrIPv6.cpp
@@ -1,4 +1,4 @@
-/* $Id: RTStrIPv6.cpp 43171 2012-09-04 17:07:27Z klaus.espenlaub@oracle.com $ */
+/* $Id: RTStrIPv6.cpp 43172 2012-09-04 17:11:51Z klaus.espenlaub@oracle.com $ */
 /** @file
  * IPRT - RTStrIPv6.
  */
@@ -568,7 +568,7 @@ static int rtStrParseAddrStr6(const char *pszAddress, size_t addressLength, char
                 }
                 else
                 {
-                    if ((pFieldEnd - pFieldStart) > fieldLength)
+                    if ((size_t)(pFieldEnd - pFieldStart) > fieldLength)
                     {
                         fieldLength = pFieldEnd - pFieldStart;
                         pFieldStartLongest = pFieldStart;
@@ -578,7 +578,7 @@ static int rtStrParseAddrStr6(const char *pszAddress, size_t addressLength, char
                     pFieldEnd = NULL;
                 }
             }
-            if (!(pGap == pszResultAddressStart && (pNow - pStart) == colons))
+            if (!(pGap == pszResultAddressStart && (size_t)(pNow - pStart) == colons))
             {
                 memset(pszResultAddress, '0', 4);
                 pMisc = pszResultAddress + 3;
